changing unit tests to integration dir
