+++ tests/integration/application/core/test_lookup_functions.py
